% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coptest.R
\name{coptest}
\alias{coptest}
\title{Two sample test for empirical copula.}
\usage{
coptest(x1, x2, nperm = 100, approx = TRUE)
}
\arguments{
\item{x1}{Numeric matrix. Samples in row and variables in column}

\item{x2}{Numeric matrix. The same with x1.}

\item{nperm}{The number of permutation.}

\item{approx}{Logical. If `approx=TRUE`, p-value will be approximated using generalized Parato distribution. Otherwise, no approximation of p-value.}
}
\value{
List of three components:
\itemize{
  \item pval - p-value
  \item stat - test statistics based on Cramér–von Mises test statistic
  \item null - null distribution from randomization procedure
}
}
\description{
Two sample test for empirical copula.
}
\examples{
data(ccrcc.pbaf) # example data from clear renal cell carcinoma(clerk et al.2019)
data(corum.hsp.pbaf)
tumor = ccrcc.pbaf$tumor # 110 samples and 10 proteins from PBAF complex
normal = ccrcc.pbaf$normal # 84 samples and 10 proteins from PBAF complex

# multivariate copula test(more than three variables)
result = coptest(tumor,normal,nperm=100,approx=FALSE)
result$pval

}
\references{
{
Yusuke MATSUI et al.(2020) RoDiCE: Robust differential protein co-expression analysis for cancer complexome (submitted).

Clerk DJ et al.(2019) Integrated Proteogenomic Characterization of Clear Cell Renal Cell Carcinoma.Cell;179(4),964-983 e931.
}
}
\author{
Yusuke MATSUI
}
