% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corum.R
\docType{data}
\name{corum}
\alias{corum}
\alias{corum.hsp}
\alias{corum.hsp.pbaf}
\alias{corum.subset}
\title{Protein membership to the protein complex}
\format{


An object of class \code{data.frame} with 30422 rows and 3 columns.

An object of class \code{data.frame} with 27 rows and 3 columns.

An object of class \code{data.frame} with 153 rows and 3 columns.
}
\source{
\url{https://cptac-data-portal.georgetown.edu}
}
\usage{
corum.hsp

corum.hsp.pbaf

corum.subset
}
\description{
Protein membership information in protein complexes, based on the database CORUM(version 3.0).
}
\details{
`corum.subset` is the example data used in vignette.
`corum.hsp` is the complex membership information of proteins restricted to "Human", and parse.corum("Human") provides the same.
`corum.hsp.pbaf` is the membership information of the PBAF complex only.
\itemize{
  \item group_id. The protein complex identifier provided as `complexID` from CORUM.
  \item desc. The protein complex name provided as `complexName` from CORUM.
  \item varname. A group of proteins that belong to a protein complex.
}
}
\examples{
data(corum.subset)
data(corum.hsp)
data(corum.hsp.pbaf)
}
\references{
Clerk DJ et al.(2019) Integrated Proteogenomic Characterization of Clear Cell Renal Cell Carcinoma.Cell;179(4),964-983 e931.
}
\keyword{datasets}
