% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netvis.R
\name{netvis}
\alias{netvis}
\title{Visualize differential protein complex co-expression network}
\usage{
netvis(obj = NULL, tbl = NULL, ref = NULL, title = NULL, p = 0.05)
}
\arguments{
\item{obj}{data.frame. Direct result object of `coptest.p`.}

\item{tbl}{data.frame. Alternative with `obj` when you manually prepare the table. This table must include variables three variables;`varname`(pairs of variable which is separated by `|`), `stat`(test statistic), `p`(p-value), and `p.adj`(adjusted p-value).}

\item{ref}{data.frame. Group information of the variables, e.g., protein complex. The table must include `group_id`(group identifier, e,g., protein complex identifier),`desc`(description of the group,e.g., protein complex name),and `varname`(variable name, e.g., proteins)}

\item{title}{character. Main title of the plot. if `ref` is provided, `title` is set to `desc`.}

\item{p}{Single numeric. Significance level for two sample test. The default is 0.05.}
}
\value{
Interactive network.
}
\description{
Create a plot of differential co-expressed proteins within a complex.
}
\examples{
data(ccrcc.pbaf) # example data from clear renal cell carcinoma(clerk et al.2019)
data(corum.hsp.pbaf)
tumor = ccrcc.pbaf$tumor # 110 samples and 10 proteins from PBAF complex
normal = ccrcc.pbaf$normal # 84 samples and 10 proteins from PBAF complex

#perform copula test for pairwise variables.
result = coptest.p(tumor,normal,nperm=100,approx=TRUE)
result$tbl

# without reference
netvis(result,title="PBAF complex")

# with reference
data(corum.hsp.pbaf) # protein complex (PBAF complex; id=149) membership from CORUM database.
netvis(result,ref=corum.hsp.pbaf)

}
\references{
{
Yusuke MATSUI et al.(2020) RoDiCE: Robust differential protein co-expression analysis for cancer complexome (submitted).
}

{
Yusuke MATSUI et al.(2020) RoDiCE: Robust differential protein co-expression analysis for cancer complexome (submitted).
}
}
\author{
Yusuke Matsui

Yusuke MATSUI
}
