% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_corum.R
\name{parse.corum}
\alias{parse.corum}
\title{Parsing protein complex information from CORUM.}
\usage{
parse.corum(organism = "Human")
}
\arguments{
\item{organism}{Single character vector. Select from: "Human",Mouse","Pig","Bovine","Rat","Mammalia","Rabbit","Dog","Hamster"  "MINK"}
}
\value{
list consists of:
\itemize{
  \item comp.tbl - data.frame. It is consisting from complexID(complex identifier given in CORUM database), group_id(complex name),varname(subunit).
  \item comp.list - list version of comp.tbl. Each list elements corresponds to a complex.
  \item comp.raw - data.frame. Raw data from CORUM database "allComplexes.txt" from http://mips.helmholtz-muenchen.de/corum/#download.
}
}
\description{
This function parse the dataset from CORUM (ver.3.0) to obtain protein membership to protein copmlexes.
}
\details{
The data is complied from CORUM The resulting output can be used for `ref` object in netvis() function.
}
\examples{
ref = parse.corum("Human")
names(ref)
}
\references{
Giurgiu, M. et al. (2019) CORUM: the comprehensive resource of mammalian protein complexes-2019. Nucleic Acids Res, 47(D1), D559–D563.
}
