% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corum.R
\docType{data}
\name{corum}
\alias{corum}
\alias{corum.raw}
\alias{corum.hsp}
\alias{corum.hsp.pbaf}
\alias{corum.subset}
\title{Protein membership to the protein complex}
\format{


An object of class \code{data.frame} with 4274 rows and 20 columns.

An object of class \code{data.frame} with 30422 rows and 3 columns.

An object of class \code{data.frame} with 27 rows and 3 columns.

An object of class \code{data.frame} with 153 rows and 3 columns.
}
\source{
\url{http://mips.helmholtz-muenchen.de/corum/}
}
\usage{
corum.raw

corum.hsp

corum.hsp.pbaf

corum.subset
}
\description{
Protein membership information in protein complexes, based on the database CORUM(version 3.0).
}
\details{
`corum.subset` is the example data used in vignette.
`corum.hsp` is the complex membership information of proteins restricted to "Human", and parse.corum("Human") provides the same.
`corum.hsp.pbaf` is the membership information of the PBAF complex only.
`corum.raw` is the raw data from CORUM.
\itemize{
  \item group_id. The protein complex identifier provided as `complexID` from CORUM.
  \item desc. The protein complex name provided as `complexName` from CORUM.
  \item varname. A group of proteins that belong to a protein complex.
}
}
\examples{
data(corum.subset)
data(corum.hsp)
data(corum.hsp.pbaf)
data(corum.raw)
}
\references{
Giurgiu, M. et al. (2019) CORUM: the comprehensive resource of mammalian protein complexes-2019. Nucleic Acids Res, 47(D1), D559–D563.
}
\keyword{datasets}
